Add specs for Package class
